backend/requirements.txt
# Heimdall Backend Dependencies

# ── Web framework ─────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# ── Config & validation ───────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# ── Security / encryption ─────────────────────────────────────────────────────
cryptography>=42.0.0         # Fernet vault encryption
python-jose[cryptography]>=3.3.0  # JWT tokens
passlib[bcrypt]>=1.7.4       # password hashing

# ── LLM providers ─────────────────────────────────────────────────────────────
anthropic>=0.40.0
httpx>=0.27.0                # OpenAI-compat, LM Studio, Ollama, DashScope

# ── Git integration ───────────────────────────────────────────────────────────
gitpython>=3.1.43
PyGithub>=2.3.0

# ── Messaging ─────────────────────────────────────────────────────────────────
python-telegram-bot>=21.0
discord.py>=2.4.0
aiosmtplib>=3.0.0
aioimaplib>=1.1.0

# ── Task scheduling ───────────────────────────────────────────────────────────
apscheduler>=3.10.0

# ── Utilities ─────────────────────────────────────────────────────────────────
anyio>=4.4.0
aiofiles>=23.2.1
jinja2>=3.1.4
